teach cli how to fetch a template from an OCI registry
